Fix crash on servers
